fix(gui): read float before int for CRF default value
EncoderOptionVal has both Int and Float fields. For libx264 CRF,
the Int field is set to -1 while the Float field holds the correct
default of 23.0. The previous code used int ?? float, so -1 won
over 23.0. Swap to float ?? int so the floating-point default is
preferred.

+++ frontend/src/lib/streamAttrs.ts
@@ -321,7 +321,7 @@ function attrsFromGraphNode(node: NodeDef, type: StreamType): Record<string, str
           const defaultRc = roles.default_rc ?? (roles.crf ? 'crf' : roles.qp ? 'qp' : 'bitrate');
           if ((defaultRc === 'crf') && roles.crf) {
             const crfOpt = findOption(info.options, roles.crf);
-            const defVal = crfOpt?.default?.int ?? crfOpt?.default?.float;
+            const defVal = crfOpt?.default?.float ?? crfOpt?.default?.int;
             const label  = roles.crf === 'cq' ? 'CQ' : roles.crf === 'global_quality' ? 'ICQ' : 'CRF';
             out['rate_control'] = defVal !== undefined
               ? `${label} ${defVal} (default)`
